Let the downloaders implement their own progress events and mark the old event as obsolete.
This solves a problem when the download is much slower than the audio conversion,
so it downloads to 50% progress and then suddenly jumps to 100% due to the audio extraction being much faster in relation to the download.

Now users of the AudioDownloader can subscribe to the DownloadProgressChanged and AudioExtractionProgressChanged events separately.