Decipher Performance Enhancement
Removed unnecessary code duplication in Decipherer.cs
Added a loop breakout when all three functions have been determined.
Rearranged the code so that operations is determined once and applied to
each encrypted signature instead. This should greatly improve the
performance of the decipher as it won't have to run though all the loops
each time to decrypt a video/url.
Renamed some variables and comments to clear up some of the trial fix I
made.